fix: correct emoji rendering and prevent row height expansion
- Update emoji unicode range detection and sequence parsing in `smart_font.cpp` to correctly handle Variation Selectors and ZWJ.
- Fix Qt font double scaling issue by using `setPixelSize` instead of `setPointSize`.
- Clamp ascent and descent of the emoji font in `qt_font_rep` to prevent unexpectedly large table row heights.
- Enhance QTMKeyboardEvent to handle 32-bit codepoints (emojis) correctly.

+++ src/Plugins/Qt/QTMKeyboardEvent.hpp
@@ -110,7 +110,7 @@ private:
     string mTexmacsKeyCombination;
     QString nss;
     unsigned int   kc;
-    unsigned short unic;
+    unsigned int   unic;
     
 
 };
